Fix #82 p.1
