+++ AdamController/Views/MainWindow.xaml
@@ -81,14 +81,6 @@
                     <iconPacks:PackIconModern Padding="0 3 0 3" Kind="ChevronRight" />
                 </Button>
 
-                <TextBlock HorizontalAlignment="Center"
-                           VerticalAlignment="Center"
-                           Margin="10 0 10 0"
-                           Text="{Binding VideoFrameRate}"
-                           Foreground="{DynamicResource MahApps.Brushes.Text}"
-                           Style="{DynamicResource MahApps.Styles.TextBlock.Watermark}"/>
-
-
             </StackPanel>
         </mah:WindowCommands>
     </mah:MetroWindow.RightWindowCommands>
